test: split catalog-ingest task tests into test_ingest_tasks.py
The wiring-gap fix series (#156/#159/#162) grew
test_scheduling_tasks.py past the 800-line audit gate (806). Move the
five dataops.ingest_* task test classes (judicial/conamer/rmf/nom/
treaty) into their own file — they're one cohesive 3-test-pattern
family — instead of allowlisting a test file.

Also exclude .claude/ (agent worktrees hold full repo copies) from the
file-size audit so local runs match CI.
